Para pruebas usuarios

+++ src/main/java/app/retos/controllers/ContactsController.java
@@ -1,6 +1,7 @@
 package app.retos.controllers;
 
 import app.retos.models.Contacts;
+import app.retos.repository.ContactsRepository;
 import app.retos.services.IContactsService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,6 +23,9 @@ public class ContactsController {
     @Autowired
     IContactsService contactsService;
 
+    @Autowired
+    ContactsRepository contactsRepository;
+
     @PostMapping("/crear/{username}")
     @ResponseStatus(code = HttpStatus.CREATED)
     public String crearContactoUsuario(@PathVariable("username") String username, @RequestBody @Validated Contacts contacts) throws IOException {
@@ -37,9 +41,12 @@ public class ContactsController {
     public String editarContacto(@PathVariable("username") String username, @PathVariable("email") String email,
                                  @PathVariable("cellPhone") String cellPhone, @RequestBody Contacts contacts) throws IOException {
         if (usersController.EmailUsernameUsuarioExiste(username)) {
-            if (contactsService.editarContactos(username, email, cellPhone, contacts))
-                return "Contacto editado correctamente";
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Error en la edición de contactos");
+            if(contactsRepository.existsByUsernameAndEmailAndCellPhone(username, email, cellPhone)){
+                if (contactsService.editarContactos(username, email, cellPhone, contacts))
+                    return "Contacto editado correctamente";
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Error en la edición de contactos");
+            }
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "El contacto no existe");
         }
         throw new ResponseStatusException(HttpStatus.NOT_FOUND, "El usuario: " + username + " no existe");
     }
@@ -49,8 +56,10 @@ public class ContactsController {
     public String eliminarContacto(@PathVariable("username") String username, @PathVariable("email") String email,
                                    @PathVariable("cellPhone") String cellPhone) throws IOException {
         if (usersController.EmailUsernameUsuarioExiste(username)) {
-            if (contactsService.eliminarContacto(username, email, cellPhone)) return "Contacto eliminado correctamente";
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Error en la edición de contactos");
+            if(contactsRepository.existsByUsernameAndEmailAndCellPhone(username, email, cellPhone)){
+                if (contactsService.eliminarContacto(username, email, cellPhone)) return "Contacto eliminado correctamente";
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Error en la eliminacion de contactos");
+            }throw new ResponseStatusException(HttpStatus.NOT_FOUND, "El contacto no existe");
         }
         throw new ResponseStatusException(HttpStatus.NOT_FOUND, "El usuario: " + username + " no existe");
     }
